+++ src/lib/camera/camera_control.py
