+++ src/lib/data_connector/data_connector.py
@@ -1,4 +1,4 @@
-from model.serialization.serialization import serializeScan, deserializeScan
+from lib.model.serialization.serialization import serializeScan, deserializeScan
 
 import blosc
 import os
