+++ src/lib/model/scan.py
@@ -1,4 +1,4 @@
-from model.sweep import Sweep
+from lib.model.sweep import Sweep
 
 
 class Scan:
